+++ NEWS.md
@@ -3,6 +3,8 @@
 * Add `get_handler()` method to `Route` (#9, @cpsievert)
 * Add `root` field to `Route` which will get appended to all paths before 
   matching to an incomming request
+* Add `remap_handlers()` to loop through all handlers and reassign them based on
+  a user provided function. (#8)
 
 # routr 0.3.0